% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box-path.R
\name{box_path}
\alias{box_path}
\title{Build paths relative to where Box is located}
\usage{
box_path(..., .box_root = NULL)
}
\arguments{
\item{...}{String elements that will be combined with the Box
root to form a path (see \code{?file.path})}

\item{.box_root}{A path to where the Box is located, if it is
installed in a non-standard place. If \code{NULL}, this utility will
also check the "BOX_PATH" environmental variable to see if it's
defined. This method is useful for containerized code.}
}
\value{
An absolute, normalized path
}
\description{
Box can be found on multiple locations on a device. This utility
searches for common locations, both on macOS and Windows. From
there, it builds paths similar to how \code{file.path()} operates.
}
\examples{
if (FALSE) {
  box_path("Box 3EA Team Folder/Data Management")
  box_path("Peru/Data/Full")
}
}
