% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-omega.R
\name{calc.omega}
\alias{calc.omega}
\title{Calculate omega squared from CFA models}
\usage{
calc.omega(loadings, resid)
}
\arguments{
\item{loadings}{Factor loadings from a CFA model}

\item{resid}{Residuals from a CFA model}
}
\value{
A single omega value
}
\description{
\code{calc.omega()}takes loadings and residuals from a CFA model
to calculate McDonald's Omega (see Revelle & Zinbarg, 2008). It is an intermediate function for
\code{get.omega.bywave()} and \code{get.omega.lg()}, and therefore shouldn't be run alone.
}
\note{
Revelle, W., & Zinbarg, R. E. (2008). Coefficients Alpha, Beta, Omega, and the glb: Comments on Sijtsma. Psychometrika, 74(1), 145. https://doi.org/10.1007/s11336-008-9102-z
}
\seealso{
\code{get.omega.bywave()} and \code{get.omega.lg()}
}
