% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-cfa-fit.R
\name{get.cfa.fit}
\alias{get.cfa.fit}
\title{Get CFA model fit}
\usage{
get.cfa.fit(model, path)
}
\arguments{
\item{model}{Mplus model name}

\item{path}{Mplus model file path}
}
\value{
A data.frame of CFA model fits
}
\description{
Get CFA model fit
}
\seealso{
\code{mplus.version.fit.short}
}
