% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-cor-bivar.R
\name{get.cor.bivar}
\alias{get.cor.bivar}
\title{Get bivariate correlation matrices from factor scores}
\usage{
get.cor.bivar(fs_data)
}
\arguments{
\item{fs_data}{A data.frame of the factor score data saved from Mplus OR raw data}
}
\value{
A data.frame of the correlation matrix.
}
\description{
Get bivariate correlation matrices from factor scores
}
\note{
Factor score data from Mplus do not have variable headers. You may need to manually add them before importing the data to R.
}
\seealso{
\code{get.cor.lg}
}
