% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-cor-lg.R
\name{get.cor.lg}
\alias{get.cor.lg}
\title{Get correlation matrices from longitudinal invariance models}
\usage{
get.cor.lg(model, path, string = ".WITH$")
}
\arguments{
\item{model}{Mplus model name}

\item{path}{Mplus file path}

\item{string}{Default is set to \code{".WITH$"} to identify rows ending with \code{WITH} in paramHeader.
No need to further specify \code{string = ".WITH$"} if used alone.}
}
\value{
A data.frame of the correlation matrix.
}
\description{
Get correlation matrices from longitudinal invariance models
}
\seealso{
\code{get.cor.bivar}
}
