% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-eigenvalue.R
\name{get.eigenvalue}
\alias{get.eigenvalue}
\title{Get eigenvalue for screeplots from EFA models}
\usage{
get.eigenvalue(model, path)
}
\arguments{
\item{model}{Mplus model name}

\item{path}{Mplus model file path}
}
\value{
A data.frame of wave tags and eigenvalues for factor n
}
\description{
Get eigenvalue for screeplots from EFA models
}
