% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-est.R
\name{get.est}
\alias{get.est}
\title{Get CFA model estimates (for plotting)}
\usage{
get.est(model, path)
}
\arguments{
\item{model}{Mplus model name}

\item{path}{Mplus model file path}
}
\value{
A data.frame of wave tags and corresponding model estimates
}
\description{
Get CFA model estimates (for plotting)
}
\note{
stdyx.standardized is used if available. Otherwise, stdy.standardized is used.
}
\seealso{
\code{get.modparam}
}
