% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-invariance-fit.R
\name{get.invariance.fit}
\alias{get.invariance.fit}
\title{Get measurement invariance model fits}
\usage{
get.invariance.fit(inv_models, path)
}
\arguments{
\item{inv_models}{A vector of Mplus measurement invariance model name, in the order of config, metric and scalar model names}

\item{path}{Mplus model file path}
}
\value{
A data.frame of measurement invariance model fits
}
\description{
Get measurement invariance model fits
}
\seealso{
\code{mplus.version.fit.long}
}
