% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-modparam.R
\name{get.modparam}
\alias{get.modparam}
\title{Get CFA model estimates (for tables)}
\usage{
get.modparam(model, path)
}
\arguments{
\item{model}{Mplus model name}

\item{path}{Mplus model file path}
}
\value{
A data.frame of wave tags, corresponding model estimates and SEs
}
\description{
Get CFA model estimates (for tables)
}
\note{
stdyx.standardized is used if available. Otherwise, stdy.standardized is used.
}
\seealso{
\code{get.est}
}
