% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-omega-bywave.R
\name{get.omega.bywave}
\alias{get.omega.bywave}
\title{Get omega squared from CFA at each wave}
\usage{
get.omega.bywave(model, path)
}
\arguments{
\item{model}{Mplus model name}

\item{path}{Mplus model file path}
}
\value{
A data.frame with two variables: \code{subscale_wave} and \code{omega_by_wave}
}
\description{
\code{get.omega.bywave} creates McDonald's omega from an Mplus CFA model on a given wave data.
\code{calc.omega()} is called to calculate omegas. \code{loadings} come from factor loadings of the CFA model,
and \code{resid} comes from r2 of the CFA model.
}
\note{
Mac users can hit \code{command + option + c} in finder to quickly get the local file path.
}
