% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-omega-lg.R
\name{get.omega.lg}
\alias{get.omega.lg}
\title{Get omega squared from longitudinal invariance models}
\usage{
get.omega.lg(model, path)
}
\arguments{
\item{model}{Mplus model name}

\item{path}{Mplus model file path}
}
\value{
A data.frame with two variables: \code{subscale_wave} and \code{omega_lg}
}
\description{
\code{get.omega.lg} creates McDonald's omega from an Mplus longitudinal invariance model.
\code{calc.omega()} is called to calculate omegas. \code{loadings} come from factor loadings in stdyx.standardized,
and \code{resid} comes from resid_var.
}
\note{
Mac users can hit \code{command + option + c} in finder to quickly get the local file path.
}
