% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-R2.R
\name{get.R2}
\alias{get.R2}
\title{Get CFA model R-squared}
\usage{
get.R2(model, path)
}
\arguments{
\item{model}{Mplus model name}

\item{path}{Mplus model file path}
}
\value{
A data.frame of wave tags, corresponding R2 estimates, SEs, p values and residual variances
}
\description{
Get CFA model R-squared
}
