% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-mplus-version-long.R
\name{mplus.version.fit.long}
\alias{mplus.version.fit.long}
\title{Check Mplus version and get model fits}
\usage{
mplus.version.fit.long(df)
}
\arguments{
\item{df}{A data.frame from readModels}
}
\value{
A data.frame of invariance model fits
"Parameters", "ChiSqM_Value", "ChiSqM_DF", "ChiSqM_PValue"
"ChiSqBaseline_Value","ChiSqBaseline_DF","ChiSqBaseline_PValue"
"ChiSqDiffTest_Value","ChiSqDiffTest_DF","ChiSqDiffTest_PValue",
"SRMR", "CFI", "TLI", "RMSEA_Estimate", "SRMR"/"WRMR", "Filename"
}
\description{
This function checks if Mplus.version >= 8, if yes, if it reports SRMR along with other statistics;
if no, it reports WRMR, which was built in replace of SRMR in earlier versions of Mplus.
In the knitted report, the short version is used for CFA model fits, and the long version is used for invariance model fits.
}
\note{
This function selects ChiSqBaseline_Value and ChiSqDiffTest_Value and their corresponding df and p value,
which are not included in \code{mplus.version.fit.short}
}
\examples{
df <- readModels(target = file.path(path,model))$summaries
mplus.version.fit.long(df)
}
\seealso{
\code{mplus.version.fit.short}
}
