% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-mplus-version-short.R
\name{mplus.version.fit.short}
\alias{mplus.version.fit.short}
\title{Check Mplus version and get model fits}
\usage{
mplus.version.fit.short(df)
}
\arguments{
\item{df}{A data.frame from readModels}
}
\value{
A data.frame of CFA model fits: "k", "ChiSq", "df", "p", "CFI", "TLI", "RMSEA", "SRMR"/"WRMR", "Filename"
}
\description{
This function checks if Mplus.version >= 8, if yes, if it reports SRMR along with other statistics;
if no, it reports WRMR, which was built in replace of SRMR in earlier versions of Mplus.
In the knitted report, the short version is used for CFA model fits, and the long version is used for invariance model fits.
}
\examples{
\dontrun{
df <- MplusAutomation::readModels(target = file.path(path,model))$summaries
mplus.version.fit.short(df)
}
}
\seealso{
\code{mplus.version.fit.long}
}
