% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrautomatr-package.R
\docType{package}
\name{mrautomatr-package}
\alias{mrautomatr}
\alias{mrautomatr-package}
\title{TIES Measurement Report Automation Project (\code{mrautomatr})}
\description{
This project is developed to generate lengthy but informative measurement reports from survey data and Mplus measurement models.
}
\section{Functions}{

\code{\link[=calc.omega]{calc.omega()}},
\code{\link[=check.mplus.version.long]{check.mplus.version.long()}}, \code{\link[=check.mplus.version.short]{check.mplus.version.short()}},
\code{\link[=get.cfa.fit]{get.cfa.fit()}}, \code{\link[=get.cor.bivar]{get.cor.bivar()}}, \code{\link[=get.cor.lg]{get.cor.lg()}}, \code{\link[=get.eigenvalue]{get.eigenvalue()}},
\code{\link[=get.est]{get.est()}}, \code{\link[=get.invariance.fit]{get.invariance.fit()}}, \code{\link[=get.modparam]{get.modparam()}}, \code{\link[=get.omega.bywave]{get.omega.bywave()}},
\code{\link[=get.omega.lg]{get.omega.lg()}}, \code{\link[=get.R2]{get.R2()}},
\code{\link[=render_report_manual]{render_report_manual()}}, \code{\link[=render_report_multiple]{render_report_multiple()}}, \code{\link[=render_report]{render_report()}}
}

\author{
\strong{Maintainer}: Zezhen Wu \email{zw1429@nyu.edu} (\href{https://orcid.org/0000-0002-5150-6736}{ORCID})

}
\keyword{internal}
