% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-report-manual.R
\name{render_report_manual}
\alias{render_report_manual}
\title{Render measurement report by manually specifying parameters}
\usage{
render_report_manual(output_file, output_dir)
}
\arguments{
\item{output_file}{The name of the output file.}

\item{output_dir}{The output directory for the rendered \code{output_file}.}
}
\value{
A docx document.
}
\description{
This function opens a Shiny app that allows you to type in global information about the measurement report
and click/unclick sections that you don't want in the report.
}
