% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-report-multiple.R
\name{render_report_multiple}
\alias{render_report_multiple}
\title{Render multiple measurement reports at once}
\usage{
render_report_multiple(
  input_dir,
  templates = list.files(input_dir, pattern = "^[a-zA-Z].*\\\\.xlsx$"),
  output_dir
)
}
\arguments{
\item{input_dir}{A string of the input directory where the template files are located}

\item{templates}{A string of prespecified .xlsx templates file names (see README.md); if not specified, all files with the .xlsx suffix will be loaded and run (excel lockfiles "~.xlsx" are excluded).}

\item{output_dir}{A string of the output directory for the rendered output_file. This allows for a choice of an alternate directory to which the output file should be written (the default output directory of that of the input file). If a path is provided with a filename in output_file the directory specified here will take precedence. Please note that any directory path provided will create any necessary directories if they do not exist.}
}
\value{
Multiple docx documents, each with the name "Report_index.docx". Index is extracted from the file name of the input templates.
}
\description{
Render multiple measurement reports at once
}
\examples{
\dontrun{
render_report_multiple(input_dir = "/Users/michaelfive/Google Drive/NYU/3EA/test",
                       templates = c("input_template_lebanon_cs.xlsx",
                                     "input_template_niger_psra.xlsx"),
                       output_dir = "/Users/michaelfive/Google Drive/NYU/3EA/test")
                      }
}
\seealso{
\code{render_report_manual}, \code{render_report_multiple}
}
