% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-report.R
\name{render_report}
\alias{render_report}
\title{Render measurement report}
\usage{
render_report(input_dir, template, index, title, output_dir)
}
\arguments{
\item{input_dir}{A string of the input directory where the template file is located}

\item{template}{A string of prespecified .xlsx template file name (see README.md)}

\item{index}{A string of the report file name}

\item{title}{A string of the report title}

\item{output_dir}{A string of the output directory for the rendered output_file. This allows for a choice of an alternate directory to which the output file should be written (the default output directory of that of the input file). If a path is provided with a filename in output_file the directory specified here will take precedence. Please note that any directory path provided will create any necessary directories if they do not exist.}
}
\value{
A docx document with the name "Report_index.docx"
}
\description{
Render measurement report
}
\examples{
\dontrun{
render_report(input_dir = "/Users/michaelfive/Google Drive/NYU/3EA/test",
              template = "input_template_lebanon_cs.xlsx",
              index = "lebanon_cs",
              title = "Lebanon Year 1 (2016-2017)",
              output_dir = "/Users/michaelfive/Google Drive/NYU/3EA/test")
              }
}
\seealso{
\code{render_report}, \code{render_report_manual}
}
