% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-report-multiple.R
\name{render_report_multiple}
\alias{render_report_multiple}
\title{Render multiple measurement reports at once}
\usage{
render_report_multiple()
}
\value{
Multiple docx documents, each with the name "Report_index.docx". Index is extracted from the file name of the input templates.
}
\description{
Render multiple measurement reports at once
}
